Swapped to float math and added gpu acceleration to pvoc - audio conversions

+++ libXCDP/include/WindowFunctions.h
@@ -8,6 +8,6 @@ struct RealFunc;
 
 namespace window {
 
-double Hann( double x );
+float Hann( float x );
 
 }}
