+++ libXCDP/src/Utility.cpp
@@ -13,52 +13,52 @@ static double pi = acos(-1);
 
 namespace xcdp {
 
-const Interpolator Interpolators::constant = []( double mix, double a, double b ) 
+const Interpolator Interpolators::constant = []( float mix, float a, float b ) 
 	{ 
 	return a;
 	};
 
-const Interpolator Interpolators::linear = []( double mix, double a, double b ) 
+const Interpolator Interpolators::linear = []( float mix, float a, float b ) 
 	{ 
 	return ( 1.0 - mix ) * a + mix * b;
 	};
 
-const Interpolator Interpolators::sine = []( double mix, double a, double b ) 
+const Interpolator Interpolators::sine = []( float mix, float a, float b ) 
 	{ 
 	return ( 1.0 - cos( pi * mix ) ) / 2.0 * ( b - a ) + a; 
 	};
 
 
 ///Code duplication?
-const Perturber Perturbers::normalDist = []( double x, double amount )
+const Perturber Perturbers::normalDist = []( float x, float amount )
 	{
 	static std::default_random_engine rng( time(nullptr) );
-	static std::normal_distribution<double> dist( 0, 1.0 );
+	static std::normal_distribution<float> dist( 0, 1.0 );
 	return x + dist(rng) * amount;
 	};
 
-const Perturber Perturbers::normalDistUp = []( double x, double amount )
+const Perturber Perturbers::normalDistUp = []( float x, float amount )
 	{
 	static std::default_random_engine rng( time(nullptr) );
-	static std::normal_distribution<double> dist( 0, 1.0 );
+	static std::normal_distribution<float> dist( 0, 1.0 );
 	return x + abs(dist(rng) * amount);
 	};
 
-const Perturber Perturbers::normalDistDown = []( double x, double amount )
+const Perturber Perturbers::normalDistDown = []( float x, float amount )
 	{
 	static std::default_random_engine rng( time(nullptr) );
-	static std::normal_distribution<double> dist( 0, 1.0 );
+	static std::normal_distribution<float> dist( 0, 1.0 );
 	return x - abs(dist(rng) * amount);
 	};
 
-const Perturber Perturbers::identity = []( double x, double amount ){ return x; };
+const Perturber Perturbers::identity = []( float x, float amount ){ return x; };
 
-std::array<uint8_t,3> HSVtoRGB( int H, double S, double V ) 
+std::array<uint8_t,3> HSVtoRGB( int H, float S, float V ) 
 	{
-	double C = S * V;
-	double X = C * (1 - abs(fmod(H / 60.0, 2) - 1));
-	double m = V - C;
-	double Rs, Gs, Bs;
+	float C = S * V;
+	float X = C * (1 - abs(fmod(H / 60.0, 2) - 1));
+	float m = V - C;
+	float Rs, Gs, Bs;
 
 		 if( H >= 0   && H < 60  ) { Rs = C; Gs = X; Bs = 0; }
 	else if( H >= 60  && H < 120 ) { Rs = X; Gs = C; Bs = 0; }
